+++ lib/moon/components/link.ex
@@ -21,13 +21,13 @@ defmodule Moon.Components.Link do
       class={
         "inline-block p-0 border-none no-underline",
         @class,
-        "hover:text-piccolo-80 focus:text-piccolo-80 active:text-piccolo-80":
+        "hover:text-piccolo focus:text-piccolo active:text-piccolo":
           !@disabled and !@optional and !@secondary,
         "cursor-pointer": !@disabled,
         "cursor-not-allowed opacity-25": @disabled,
         "text-bulma hover:underline focus:underline active:underline": !@secondary && !@optional,
         "text-trunks": @secondary,
-        "text-trunks hover:text-piccolo-80 focus:text-bulma active:text-bulma": @optional
+        "text-trunks hover:text-piccolo focus:text-bulma active:text-bulma": @optional
       }
       data-phx-link="patch"
       data-phx-link-state="push"