+++ lib/moon_web/pages/components/text_input/text_input_page.ex
@@ -121,6 +121,9 @@ defmodule MoonWeb.Pages.Components.TextInputPage do
     <Page {=@theme_name} {=@active_page} {=@breadcrumbs} {=@direction}>
       <TopToDown>
         <ComponentPageDescription title="Text Input" />
+        <p>
+          Indeed, forms require input from all users, including astronauts. Text input fields help to gather both short and long-form information. The size of the text field should correspond to the length of information that's required.</p>
+        <p>Text input fields can contain various rules to ensure the right information is obtained from users. Users should also be given prompts to guide them on what's needed, or if they've made an error while doing so.</p>
 
         <ExampleAndCode id="input_020" title="Sizes">
           <:example>
@@ -340,19 +343,9 @@ defmodule MoonWeb.Pages.Components.TextInputPage do
         <ExampleAndCode id="input_004" title="Search">
           <:example>
             <LeftToRight class="justify-between w-full items-end">
-              <TextInput
-                type="email"
-                size="md"
-                label="Medium (Default: md)"
-                placeholder="e.g. jake.weary@sportsbet.io"
-              />
-              <TextInput type="email" size="lg" label="Large (lg)" placeholder="e.g. jake.weary@sportsbet.io" />
-              <TextInput
-                type="email"
-                size="xl"
-                label="Xlarge (xl)"
-                placeholder="e.g. jake.weary@sportsbet.io"
-              />
+              <TextInput type="search" size="md" label="Medium (Default: md)" placeholder="e.g. happiness" />
+              <TextInput type="search" size="lg" label="Large (lg)" placeholder="e.g. health" />
+              <TextInput type="search" size="xl" label="Xlarge (xl)" placeholder="e.g. money" />
             </LeftToRight>
           </:example>
 
@@ -765,22 +758,22 @@ defmodule MoonWeb.Pages.Components.TextInputPage do
   def text_input_004_code do
     """
     <TextInput
-      type="email"
+      type="search"
       size="md"
       label="Medium (Default: md)"
-      placeholder="e.g. jake.weary@sportsbet.io"
+      placeholder="e.g. happiness"
     />
     <TextInput
-      type="email"
+      type="search"
       size="lg"
       label="Large (lg)"
-      placeholder="e.g. jake.weary@sportsbet.io"
+      placeholder="e.g. health"
     />
     <TextInput
-      type="email"
+      type="search"
       size="xl"
       label="Xlarge (xl)"
-      placeholder="e.g. jake.weary@sportsbet.io"
+      placeholder="e.g. money"
     />
     """
   end