+++ lib/moon_web/pages/design/menu_item_page.ex
@@ -1,8 +1,8 @@
-defmodule MoonWeb.Pages.Components.MenuItemPage do
+defmodule MoonWeb.Pages.Design.MenuItemPage do
   @moduledoc false
   use MoonWeb, :live_view
 
-  alias Moon.Components.MenuItem
+  alias Moon.Design.MenuItem
   alias Moon.Components.Lego
   alias MoonWeb.Components.ExampleAndCode
   alias MoonWeb.Components.Page
@@ -27,7 +27,7 @@ defmodule MoonWeb.Pages.Components.MenuItemPage do
         name: "Components"
       },
       %{
-        to: "/components/menu",
+        to: "/components/v2/menu_item",
         name: "MenuItem"
       }
     ]
@@ -38,70 +38,70 @@ defmodule MoonWeb.Pages.Components.MenuItemPage do
       %{
         :name => 'title',
         :type => 'string',
-        :required => 'false',
+        :required => 'No',
         :default => '-',
         :description => 'Title of item, only if no slot is given'
       },
       %{
         :name => 'text',
         :type => 'string',
-        :required => 'false',
+        :required => 'No',
         :default => '-',
         :description => 'Some hint to be shown under the title'
       },
       %{
         :name => 'is_active',
         :type => 'boolean',
-        :required => 'false',
+        :required => 'No',
         :default => 'false',
         :description => 'To mark menu item as an active (bg only)'
       },
       %{
         :name => 'is_selected',
         :type => 'boolean',
-        :required => 'false',
+        :required => 'No',
         :default => 'false',
         :description => 'Selected/checked/expanded if role is given'
       },
       %{
         :name => 'role',
         :type => 'checkbox | switch | radio',
-        :required => 'false',
+        :required => 'No',
         :default => '-',
         :description => 'Role to MenuItem acts as'
       },
       %{
         :name => 'on_click',
         :type => 'event',
-        :required => 'false',
+        :required => 'No',
         :default => '-',
         :description => 'Role-depending event handler'
       },
       %{
         :name => 'width',
         :type => 'css_class',
-        :required => 'false',
+        :required => 'No',
         :default => 'w-full',
         :description => 'Width of the item'
       },
       %{
         :name => 'class',
         :type => 'css_class',
-        :required => 'false',
+        :required => 'No',
         :default => '-',
         :description => 'Additional CSS class(es) for the item'
       },
       %{
         :name => 'as',
         :type => 'a | button',
-        :required => 'false',
+        :required => 'No',
         :default => 'button',
         :description => 'HTML tag to be used to render item'
       },
       %{
         :name => 'attrs',
         :type => 'Map | List',
-        :required => 'false',
+        :required => 'No',
         :default => '%{}',
         :description => 'HTML attributes to be set for tag given by as'
       }
@@ -450,7 +450,7 @@ defmodule MoonWeb.Pages.Components.MenuItemPage do
         <:code>{menu_item_8_code()}</:code>
       </ExampleAndCode>
 
-      <PropsTable data={@props_info_array} />
+      <PropsTable title="MeniItem props" data={@props_info_array} />
     </Page>
     """
   end